# Speed up `_notify_conda_outdated` due to expensive PrefixData call.