--- a/conda_libmamba_solver/solver.py
+++ b/conda_libmamba_solver/solver.py
@@ -1007,11 +1007,13 @@ class LibMambaSolver(Solver):
         # manually check base prefix since `PrefixData(...).get("conda", None) is expensive
         # once prefix data is lazy this might be a different situation
         current_conda_prefix_rec = None
-        for filename in os.scandir(os.path.join(context.conda_prefix, "conda-meta")):
-            if filename.name.endswith(".json") and filename.name.rsplit("-", 2)[0] == "conda":
-                with open(filename.path) as f:
-                    current_conda_prefix_rec = PrefixRecord(**json.load(f))
-                    break
+        conda_meta_prefix_directory = os.path.join(context.conda_prefix, "conda-meta")
+        if os.path.lexists(conda_meta_prefix_directory):
+            for filename in os.scandir(conda_meta_prefix_directory):
+                if filename.name.endswith(".json") and filename.name.rsplit("-", 2)[0] == "conda":
+                    with open(filename.path) as f:
+                        current_conda_prefix_rec = PrefixRecord(**json.load(f))
+                        break
         else:
             # We are checking whether conda can be found in the environment conda is
             # running from. Unless something is really wrong, this should never happen.