--- a/conda_libmamba_solver/solver.py
+++ b/conda_libmamba_solver/solver.py
@@ -11,6 +11,7 @@ import logging
 import os
 import re
 import sys
+from contextlib import suppress
 from collections import defaultdict
 from functools import lru_cache
 from inspect import stack
@@ -1008,12 +1009,13 @@ class LibMambaSolver(Solver):
         # once prefix data is lazy this might be a different situation
         current_conda_prefix_rec = None
         conda_meta_prefix_directory = os.path.join(context.conda_prefix, "conda-meta")
-        if os.path.lexists(conda_meta_prefix_directory):
-            for filename in os.scandir(conda_meta_prefix_directory):
-                if filename.name.endswith(".json") and filename.name.rsplit("-", 2)[0] == "conda":
-                    with open(filename.path) as f:
-                        current_conda_prefix_rec = PrefixRecord(**json.load(f))
-                        break
+        with suppress(OSError):
+            if os.path.lexists(conda_meta_prefix_directory):
+                for filename in os.scandir(conda_meta_prefix_directory):
+                    if filename.name.endswith(".json") and filename.name.rsplit("-", 2)[0] == "conda":
+                        with open(filename.path) as f:
+                            current_conda_prefix_rec = PrefixRecord(**json.load(f))
+                            break
         if not current_conda_prefix_rec:
             # We are checking whether conda can be found in the environment conda is
             # running from. Unless something is really wrong, this should never happen.