--- a/news/298-faster-outdated-notification
+++ b/news/298-faster-outdated-notification
@@ -1,0 +1,19 @@
+### Enhancements
+
+* Increase performance of `notify_conda_outdated` logic. (#298)
+
+### Bug fixes
+
+* <news item>
+
+### Deprecations
+
+* <news item>
+
+### Docs
+
+* <news item>
+
+### Other
+
+* <news item>