### Description

Profiling conda-libmamba-solver after the changes that @jaimergp made in https://github.com/conda/conda-libmamba-solver/pull/295 I got the following results. Without using cProfile (since profiling slows things down) 5.05s on my hardware (with nvme drive and repodata cached)

`time conda create --solver libmamba --dry-run -n test numpy -c conda-forge`

Here is the profile data collected with `snakeviz`

![image](https://github.com/conda/conda-libmamba-solver/assets/1740337/2c76a069-661c-440c-b040-28c8d2314d1d)

Notice that `_notify_conda_outdated` is taking 2s/5.41s total and it approximately 1/3 of the time it takes for the command to run. When looking at the profiling data `PrefixData(..).get('conda', None)` is expensive at `0.08ms` for each call to `_load_single_record(...)` in total 222 calls were made totalling those 2s. All just to fetch the `conda` PrefixRecord.

I explored two approaches:
 - Make PrefixData(...).get('conda', None)` lazy so it doesn't actually read every record
 - Manually read the conda prefixrecord

I worked toward implementing both and things looked fine when making prefix data lazy BUT then implementing `PrefixData._load_site_packages` as lazy was not simple and didn't look like it would actually speed things up once this function is implemented. This is due to `_load_site_packages` calling `self.iter_records()` https://github.com/conda/conda/blob/main/conda/core/prefix_data.py#L278 which in effect makes all the work to make prefixdata lazy worthless. So we are better off implementing our own.

So with this implementation we get `4.29s` for the runtime about a 15% speedup (with slower drives the speedup will be more). With the following profile

![image](https://github.com/conda/conda-libmamba-solver/assets/1740337/b236db26-9fd8-4069-ac89-3cac3b2e68ec)

Out of curiousity I compared this with `mamba` (which uses PrefixData so we would expect it to be slower). I ensuring that caching is used `5.98s`. Then I compared with micromamba getting around `2.6s`. So I would estimate that conda's overhead compared with a compiled implementation is around 1-2s as of now using conda-libmamba-solver.

### Checklist - did you ...

- [ ] Add / update necessary tests?
- [ ] Add / update outdated documentation?