@costrouc 2023-09-27
Made a fix to check that the conda-meta directory actually exists forgot that part of the `PrefixData.load(..)` implementation.