@jaimergp 2023-09-28
Errors are either expected (#296) or connection errors. This is good to go.